slack_bot/pyproject.toml
[tool.poetry]
name = "hm_slack_bot"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9.4"
httpx = "^0.18.2"
Flask = "^2.0.1"
bugsnag = "^4.1.0"
psycopg2-binary = "^2.9.1"
gunicorn = "^20.1.0"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
neovim = "^0.3.1"
pynvim = "^0.4.2"
jedi = "^0.18.0"
isort = "^5.7.0"
pylint = "^2.6.0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
